overlay: actually disable time_daemon, for good
At least on Focal, time_daemon seems to still be able to run despite
being disabled in `init.qcom.usb.rc`. I'm not sure why, but it's
probably related to the fact that this service is also defined in
hw/init.qcom.rc and init.time_daemon.rc.

So, takes another step and overlay the time_daemon executable itself
which prevents it from running at all. As this will cause Android init
to restart the service every so often, also add a device-hacks snippet
to stop that.

Fixes: https://gitlab.com/ubports/porting/community-ports/android11/fairphone-4/fairphone-fp4/-/issues/61